requirements.txt
django==5.2.7
djangorestframework
Pillow
reportlab
